+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/features/JungleStructuresVines.java
@@ -4,9 +4,11 @@ import com.google.common.collect.Sets;
 import com.telepathicgrunt.repurposedstructures.modinit.RSFeatures;
 import com.telepathicgrunt.repurposedstructures.modinit.RSStructures;
 import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.ChunkSectionPos;
+import net.minecraft.util.math.Direction;
 import net.minecraft.world.StructureWorldAccess;
 import net.minecraft.world.gen.chunk.ChunkGenerator;
 import net.minecraft.world.gen.feature.DefaultFeatureConfig;
@@ -51,6 +53,16 @@ public class JungleStructuresVines extends Feature<DefaultFeatureConfig> {
             if (world.getStructures(ChunkSectionPos.from(position), RSStructures.JUNGLE_FORTRESS).findAny().isPresent()) {
                 world.setBlockState(position, Blocks.AIR.getDefaultState(), 3);
                 RSFeatures.SHORT_VINES.generate(world, chunkGenerator, random, position, DefaultFeatureConfig.DEFAULT);
+
+                //make sure we dont cause floating vine by scheduling a tick update so vine breaks itself
+                for(Direction facing : Direction.Type.HORIZONTAL){
+                    BlockPos offset = position.offset(facing);
+                    BlockState state = world.getBlockState(offset);
+                    if(!state.canPlaceAt(world, offset)){
+                        world.getBlockTickScheduler().schedule(offset, state.getBlock(), 0);
+                    }
+                }
+
                 return true;
             }
         }